fix updateCandidateSubmissionService testcase fix after hook

+++ tests/assessments-service.integration.test.ts
@@ -528,27 +528,16 @@ describe('Integration test: Assessment', () => {
 
     afterAll(async () => {
       await deleteAssessmentService({assessmentId: createdAssessment.id})
-      await prisma.quiz.deleteMany({
-        where: {
-          id: {
-            in: quizIds,
-          },
-        },
+      await prisma.quizPointCollection.deleteMany({
+        where: {quizId: {in: quizIds}},
       })
-      await prisma.codeLanguage.deleteMany({
-        where: {
-          id: {
-            in: codeLanguageIds,
-          },
-        },
+      await prisma.submissionPoint.deleteMany({
+        where: {userId: {in: userIds}},
       })
-      await prisma.user.deleteMany({
-        where: {
-          id: {
-            in: userIds,
-          },
-        },
-      })
+      await prisma.submission.deleteMany({where: {quizId: {in: quizIds}}})
+      await prisma.quiz.deleteMany({where: {id: {in: quizIds}}})
+      await prisma.codeLanguage.deleteMany({where: {id: {in: codeLanguageIds}}})
+      await prisma.user.deleteMany({where: {id: {in: userIds}}})
     })
 
     test('it should update candidate submission', async () => {
